refactor(worker): extract worker role crate
